--- user_service/pagination.py
@@ -1,1 +1,0 @@
-from course_service.pagination import CustomPagination
